linear backend: layers/ -> operators/, factory -> training.Linear
Relocate the linear-layer impls from pithtrain/layers/ to
pithtrain/operators/ (linear.py, grouped_linear.py) and select them via the
training context instead of the layers/factory.py global. Models build
linears through training.Linear / training.GroupedLinear; setup_model binds
FP8 vs BF16 from cfg.fp8 (bool, replacing fp8_training). FP8WeightCacheControl
loses its `enabled` flag (now always active) and renames _version->version,
clear_caches->clear. Class renames GroupLinear->GroupedLinear and
FP8GroupLinear->FP8GroupedLinear (and the *Func autograd helpers).

Deleting factory.py drops ModelImplMode: fp8_training -> training.fp8, and
use_reference_fwd (debug toggle, default False) is removed -- the
`if not use_reference_fwd:` epilog detach in the 4 models becomes
unconditional and modeling.py's reference-fwd short-circuit is dropped
(reference_forward stays as a method for tests). Includes the fp8 dgrad fix
(Triton tl.trans -> transposed-address store) in the relocated kernel.

Gate is BF16 (configs fp8=False), so the FP8 path incl. the dgrad fix is not
exercised by the lb run; lands on prior validation. tests/test_muon_fsdp.py
still imports the long-removed DistributedCtx (unmaintained; broken on the
refactor tip too) -- deferred.
