+++ pithtrain/contexts/training.py
@@ -1,13 +1,11 @@
 """
 Training runtime state, populated once at startup.
 
-Import the module and read fields in-line, not the names. The runtime fields
-(``dataset``/``model``/``optimizers``/``schedulers``/``step``) do not exist until setup assigns
-them, so reading them early raises AttributeError. The linear backend defaults to BF16 and is
-switched to FP8 by ``setup_model``.
+Import the module and read fields in-line, not the names: a field does not exist until setup
+assigns it, so importing it up front fails and reading it early raises AttributeError.
 
-    from pithtrain.contexts import training
-    self.gate_proj = training.linear_cls(hidden_size, intermediate_size, bias=False)
+from pithtrain.contexts import training
+self.gate_proj = training.Linear(hidden_size, intermediate_size, bias=False)
 """
 
 import torch.nn as nn
@@ -15,17 +13,16 @@ from torch.optim import Optimizer
 from torch.optim.lr_scheduler import LRScheduler
 
 from pithtrain.dualpipe import DualPipeV
-from pithtrain.layers.group_linear import GroupLinear
 from pithtrain.modules.dataset import ConcatDataset
+from pithtrain.operators import grouped_linear
 
-# Linear backend: BF16 by default; setup_model selects FP8 (DeepGEMM) when enabled.
 fp8: bool = False
-linear_cls: type = nn.Linear
-group_linear_cls: type = GroupLinear
+Linear: type[nn.Linear] = nn.Linear
+GroupedLinear: type[grouped_linear.GroupedLinear | grouped_linear.FP8GroupedLinear] = grouped_linear.GroupedLinear
 
 # Populated once at startup by the training setup.
-dataset: ConcatDataset
 model: DualPipeV
+dataset: ConcatDataset
 optimizers: tuple[Optimizer, ...]
 schedulers: tuple[LRScheduler, ...]
 step: int
